move Group Dto to QueryResult
